fix: Update priceAtSale field in createOrderSchema to allow non-integer values

+++ src/validators/index.ts
@@ -23,7 +23,7 @@ export const createOrderSchema = z.object({
       productId: z.number().int().positive('Product ID must be positive'),
       amount: z.number().int().positive('Amount must be positive'),
       notes: z.string().optional(),
-      priceAtSale: z.number().int().optional(),
+      priceAtSale: z.number().optional(),
     })
   ).min(1, 'At least one item is required'),
 });
